Hibakezelés, lombok használata

+++ main/src/main/java/com/inerview/main/controller/StudentController.java
@@ -2,6 +2,7 @@ package com.inerview.main.controller;
 
 import com.inerview.main.model.Student;
 import com.inerview.main.services.StudentService;
+import org.apache.commons.validator.routines.EmailValidator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -18,35 +19,45 @@ public class StudentController {
     @Autowired
     private StudentService studentService;
 
+    //az összes tanuló listázása
     @GetMapping("/students")
     public List<Student> getStudents(){
         List<Student> lista = studentService.getAllStudent();
         return lista;
     }
 
+    //a megadott id-vel rendelkező tanuló módosítása
     @PutMapping("/students/{studentId}")
-    public Student updateStudent(@PathVariable UUID studentId, @RequestBody Student student){
+    public ResponseEntity<String> updateStudent(@PathVariable UUID studentId, @RequestBody Student student){
         Student modifiable = studentService.getStudentById(studentId);
-        if(modifiable != null){
+        boolean isGoodEmail = EmailValidator.getInstance().isValid(student.getEmail());
+        if(modifiable != null && isGoodEmail){
             modifiable.setName(student.getName());
             modifiable.setEmail(student.getEmail());
             modifiable.setId(studentId);
             studentService.deleteById(studentId);
             studentService.save(modifiable);
+            return ResponseEntity.status(HttpStatus.OK).body("Modification was success.");
         }
         else{
-            //hiba
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Id or the email was incorrect.");
         }
-        return modifiable;
     }
 
+    //új tanuló beillesztése
     @PostMapping("/students")
-    public Student addStudent(@RequestBody Student student){
+    public ResponseEntity<String> addStudent(@RequestBody Student student){
         student.setId(UUID.randomUUID());
-        studentService.save(student);
-        return student;
+        Student returnValue = studentService.save(student);
+        if(returnValue != null){
+            return ResponseEntity.status(HttpStatus.OK).body("New student inserted.");
+        }
+        else{
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Can not insert student, because email is invalid.");
+        }
     }
 
+    //tanuló törlése
     @DeleteMapping("/students/{studentId}")
     public ResponseEntity<String> deleteStudent(@PathVariable UUID studentId){
         Student deletedStudent = studentService.getStudentById(studentId);
@@ -55,8 +66,7 @@ public class StudentController {
             return ResponseEntity.status(HttpStatus.OK).body("Requested student deleted.");
         }
         else{
-
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Requested id not found.");
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Can not insert student, because email is invalid.");
         }
     }
